+++ spotdl_gui/newGui.py
@@ -4,8 +4,8 @@ import qdarktheme
 from assets import resource
 from PySide6 import QtWidgets
 from PySide6.QtWidgets import QDialog, QWidget
-from views.About import Ui_About
-from views.MainWindow import Ui_MainWindow
+from views.about import Ui_About
+from views.mainwindow import Ui_MainWindow
 
 
 class AboutDialog(QtWidgets.QDialog):
