+++ spotdl_gui/views/mainwindow.py
